huge style refactor
* no more styles.css catchall. split up into several .css files...
* also deleted old Utils.css, which was copy-pasted from parametric-protocols project
* globals.css imported into App
* colors.css & mixins.css
* individual ComponentName.css for all components that were using
* removed old styles from pre-svg deckmap
