+++ protocol-designer/src/components/Deck.css
@@ -1,0 +1,10 @@
+.deck {
+  max-height: 90vh;
+}
+
+.deck_outline {
+  width: 100%;
+  height: 100%;
+  fill: transparent;
+  stroke: black;
+}
