+++ protocol-designer/src/components/Well.js
@@ -52,50 +52,3 @@ export default function Well ({
       r={wellLocation.diameter / 2}
     />
 }
-
-// TODO: remove. No longer used in SVG deckmap.
-// import styles from '../css/style.css' // TODO use own styles
-//
-// Well.propTypes = {
-//   x: PropTypes.number.isRequired,
-//   y: PropTypes.number.isRequired,
-//   wellContent: PropTypes.shape({
-//     number: PropTypes.number,
-//     selected: PropTypes.bool,
-//     preselected: PropTypes.bool,
-//     highlighted: PropTypes.bool,
-//     hovered: PropTypes.bool,
-//     groupId: PropTypes.string
-//   }).isRequired
-// }
-//
-// export default function Well ({x, y, wellContent, selectable, ...otherProps}) {
-//   const { preselected, selected, highlighted, hovered, groupId } = wellContent
-//   const isFilled = (groupId !== null && groupId !== undefined)
-//   return (
-//     <div
-//       className={cx(
-//         styles.well_round,
-//         {[styles.selected]:
-//           selected,
-//           [styles.preselected]: preselected,
-//           [styles.highlighted]: highlighted,
-//           [SELECTABLE_WELL_CLASS]: selectable && !isFilled}
-//       )}
-//       data-well-x={x}
-//       data-well-y={y}
-//       style={{
-//         '--well-selection-color': selected
-//           ? 'blue' // <- set color swatch for ingredient here
-//           : (preselected ? 'lightcyan' : 'transparent'),
-//         '--well-fill-color': isFilled
-//           ? swatchColors(parseInt(groupId, 10))
-//           : 'transparent'  // <- set well fill color here (probably, add it in wellContents)
-//       }}
-//       {...otherProps}
-//       >
-//       {/* TODO: hovered prop */}
-//       {hovered && isFilled && <WellToolTip wellContent={wellContent} />}
-//       <div className={styles.inner_well} /></div>
-//   )
-// }
