+++ protocol-designer/src/css/globals.css
@@ -1,0 +1,44 @@
+@value blue, lightGrey, medGrey, darkGrey, plateBg from '../css/colors.css';
+@value fontSizeHuge, fontSizeHeader, fontSizeBody2, fontSizeBody1, fontSizeCaption from '../css/fontSizes.css';
+
+/* NOTE Ian 2017-12-06 these globals aren't really carefully designed at the moment */
+
+html,
+body {
+  margin: 0;
+  padding: 0;
+  height: 100%;
+}
+
+body {
+  font-family: 'Open Sans', sans-serif;
+}
+
+ul {
+  list-style-type: none;
+}
+
+h1 {
+  font-size: fontSizeHuge;
+}
+
+h2 {
+  color: gray;
+  font-size: fontSizeHeader;
+  text-align: center;
+}
+
+label {
+  display: inline-block;
+  text-transform: uppercase;
+}
+
+input,
+textarea {
+  border: 1px solid gray;
+}
+
+textarea {
+  /* textarea not scrollable by default */
+  resize: none;
+}
