+++ protocol-designer/src/css/mixins.css
@@ -1,0 +1,14 @@
+.disable_select {
+  -webkit-user-select: none;
+  -moz-user-select: none;
+  -ms-user-select: none;
+  user-select: none;
+}
+
+.close {
+  /* TODO: rethink this */
+  position: absolute;
+  right: 0;
+  top: 0;
+  cursor: pointer;
+}
