--- protocol-designer/src/css/style.css
@@ -1,452 +1,0 @@
-@value slotWidth: 85.47;
-@value slotHeight: 127.76;
-@value slotScale: 0.125rem;
-@value slotGap: 10px;
-
-/* COLORS */
-@value lightgray: #d0d0d0;
-@value gray: #9a9a9a;
-@value darkgray: #5f5f5f;
-@value overlayColor: rgba(0, 0, 0, 0.7);
-
-/* GLOBAL STYLES! */
-
-html,
-body {
-  margin: 0;
-  padding: 0;
-  height: 100%;
-}
-
-body {
-  font-family: 'Open Sans', sans-serif;
-}
-
-ul {
-  list-style-type: none;
-}
-
-div {
-  border-radius: 5px;
-}
-
-h2 {
-  color: gray;
-  text-align: center;
-}
-
-label {
-  display: inline-block;
-  text-transform: uppercase;
-}
-
-input,
-textarea {
-  border: 1px solid gray;
-}
-
-button,
-.btn {
-  font-size: 24pt;
-  color: lightgray;
-  text-transform: uppercase;
-  border: 3px solid lightgray;
-  padding: 3px;
-  border-radius: 3px;
-  background-color: transparent;
-  cursor: pointer;
-}
-
-button:hover,
-.btn:hover {
-  background-color: lightgray;
-  color: white;
-}
-
-button.disabled,
-.btn.disabled {
-  background-color: #eee;
-  color: lightgray;
-  cursor: default;
-}
-
-textarea {
-  resize: none;
-}
-
-/* MIXINS */
-
-.disable_select {
-  -webkit-user-select: none;
-  -moz-user-select: none;
-  -ms-user-select: none;
-  user-select: none;
-}
-
-.overlay {
-  position: absolute;
-  top: 0;
-  bottom: 0;
-  left: 0;
-  right: 0;
-}
-
-/* General */
-
-.circle {
-  display: inline-block;
-  border-radius: 1000px;
-  width: 2rem;
-  height: 2rem;
-  margin-left: 0.8rem;
-  font-size: 1rem;
-  line-height: 2rem;
-}
-
-/* WELLS */
-
-.well_round {
-  display: flex;
-  position: relative;
-  box-sizing: border-box;
-  border-radius: 1000px;
-  border: 2px solid transparent;
-  padding: 2px;
-
-  /* margin: 2px; */
-
-  /* padding-left: 100%; */
-  height: 100%;
-  width: 100%;
-  flex: 1;
-}
-
-.well_round.highlighted {
-  border: 2px solid yellow; /* highlight color */
-}
-
-.well_round.preselected {
-  border: 2px solid powderblue; /* highlight color */
-}
-
-.well_round .inner_well {
-  display: flex;
-  box-sizing: border-box;
-  background-color: var(--well-fill-color);
-  border: 2px solid black;
-  content: attr(data-well-number);
-  font-size: 10.5rem;
-  color: white;
-  justify-content: center;
-  align-items: center;
-  border-radius: 1000px;
-  height: 100%;
-  width: 100%;
-  padding: 5%;
-}
-
-.well_round.selected .inner_well {
-  border: 4px solid var(--well-selection-color);
-}
-
-/* DECK */
-
-.deck {
-  border: 3px solid darkgray;
-  display: grid;
-  grid-gap: slotGap;
-  padding: slotGap;
-  margin: auto;
-  grid-template-columns: repeat(5, calc(slotWidth * slotScale));
-  width: calc((slotWidth * slotScale * 5) + (slotGap * 4));
-}
-
-.deck_slot {
-  height: calc(slotHeight * slotScale);
-  background-color: lightgray;
-  position: relative;
-  text-align: center;
-  composes: disable_select;
-}
-
-.deck_slot img {
-  height: calc(slotHeight * slotScale);
-  width: calc(slotWidth * slotScale);
-}
-
-.deck_slot > label {
-  font-size: 30pt;
-  color: white;
-  position: inherit;
-  top: 30pt;
-}
-
-.container_overlay {
-  composes: overlay;
-  z-index: 100;
-  color: lightgray;
-  fill: lightgray;
-}
-
-.container_overlay label {
-  color: white;
-  text-transform: none;
-}
-
-.container_overlay > * {
-  visibility: hidden;
-  margin: 1rem;
-}
-
-.container_overlay:hover {
-  background-color: overlayColor;
-}
-
-.container_overlay:hover > * {
-  visibility: visible;
-  cursor: pointer;
-}
-
-.container_overlay_add_ingred:hover,
-.container_overlay_remove:hover,
-.container_overlay_copy:hover {
-  font-weight: bold;
-  color: white;
-  fill: white;
-}
-
-.container_overlay_add_ingred::after {
-  font-size: 30pt;
-  display: block;
-  content: "+";
-}
-
-.container_overlay_remove::after {
-  font-size: 20pt;
-  display: block;
-  content: "✕";
-}
-
-/* -------------------- */
-
-.deck_slot .add_labware {
-  display: none;
-}
-
-.deck_slot:hover .add_labware {
-  display: inherit;
-}
-
-.add_labware {
-  /*
-  position: absolute;
-  left: 2px;
-  right: 2px;
-  top: 2px;
-  bottom: 2px;
-  padding-top: 50%;
-  font-size: 12pt;
-  */
-  width: 100%;
-  height: 100%;
-  fill: gray;
-  color: black;
-  margin: 10px;
-  cursor: pointer;
-  text-transform: uppercase;
-  composes: disable_select;
-}
-
-.add_labware:hover {
-  font-weight: bold;
-}
-
-.add_labware::after {
-  font-size: 20pt;
-  display: block;
-  content: "+";
-}
-
-/* -------------------- */
-
-.labware_dropdown {
-  position: absolute;
-  top: 0;
-  text-align: left;
-  font-size: 10pt;
-  width: calc(slotWidth * slotScale * 1.5);
-  background-color: lightgray;
-  border: 3px solid darkgray;
-  z-index: 500;
-  color: darkgray;
-  composes: disable_select;
-}
-
-.labware_dropdown label {
-  text-transform: none;
-  font-size: 12pt;
-  font-weight: bold;
-}
-
-.labware_dropdown > ul {
-  background-color: white;
-
-  /* Overflowing, even just y, messes with tooltips... */
-
-  /* overflow-y: scroll; */
-}
-
-/* .labware_dropdown > ul li ul li {
-  color: gray;
-} */
-
-.labware_dropdown > ul li ul li:hover {
-  font-weight: bold;
-}
-
-/* TODO: can this be a mixin instead? */
-
-/* NOTE: this is only used by LabwareDropdown */
-:global(.close) {
-  position: absolute;
-  right: 0;
-  top: 0;
-  cursor: pointer;
-}
-
-/* ----- */
-
-.ingredient_modal {
-  position: absolute;
-  left: 8%;
-  right: 8%;
-  top: 0;
-  bottom: 18%;
-  background-color: rgba(100, 100, 100, 0.8);
-  z-index: 1000;
-  border-radius: 0;
-  border: 3px solid gray;
-}
-
-.ingredient_modal .top_bar {
-  color: white;
-  font-size: 14pt;
-  font-weight: bold;
-  text-transform: uppercase;
-  margin-right: 10pt;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.ingredient_modal .info {
-  margin-left: 1em;
-}
-
-.ingredient_modal .close {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: flex-end;
-}
-
-.ingredient_modal .close p {
-  margin-right: 10pt;
-}
-
-.ingredient_modal .close:hover {
-  /* color: lightgray; */
-  -webkit-filter: brightness(0.75);
-  filter: brightness(0.75);
-  cursor: pointer;
-}
-
-@value ingredient_panel_sidebar_width: 40%;
-
-.ingredient_panel_sidebar {
-  float: left;
-  display: inline-block;
-  position: relative;
-  height: 100%;
-  width: ingredient_panel_sidebar_width;
-  border-right: 3px solid gray;
-}
-
-.ingredient_panel_content {
-  border-radius: 0;
-  position: relative;
-  display: inline-block;
-  height: 100%;
-  width: calc(100% - ingredient_panel_sidebar_width - 3px);
-}
-
-.container_detail {
-  margin: 0 2%;
-}
-
-.container_detail::after {
-  /* Horizontal line */
-  content: ' ';
-  display: block;
-  background-color: lightgray;
-  width: 70%;
-  height: 3px;
-  margin: 30px auto;
-}
-
-/* ----- */
-
-.labware_list_item {
-  position: relative;
-  width: 100%;
-}
-
-.labware_list_item:hover::after {
-  content: var(--image-url);
-  background-size: 80px 60px;
-  border: 2px solid darkgray;
-  position: absolute;
-
-  /* Images too large, and need to be cropped from left */
-  zoom: 0.5;
-  clip-path: inset(0 0 0 360px);
-  top: 0;
-
-  /* left: calc(slotWidth * slotScale * 1); */
-  left: 0;
-
-  /* width: 200px;
-  height: 200px; */
-  z-index: 5000;
-}
-
-/* shows container name on deck */
-.name_overlay {
-  font-size: 0.75rem;
-  fill: white;
-}
-
-.name_overlay .container_name {
-  font-size: 1rem;
-}
-
-.deck_slot:hover .name_overlay {
-  display: none;
-}
-
-.container_overlay_name_it {
-  composes: overlay;
-  background-color: overlayColor;
-  z-index: 400;
-  color: lightgray;
-}
-
-.container_overlay_name_it > * {
-  margin: 1rem;
-}
-
-.container_overlay_name_it input {
-  font-size: 1.5rem;
-  width: 80%;
-  margin: 0;
-}
